Day 9, part one, WIP.

+++ src/main/kotlin/adventofcode/day07/Day07.kt
@@ -11,20 +11,20 @@ fun main() {
 
 // --- Types ---
 
-typealias Settings = List<Int>
+typealias Settings = List<Long>
 
 data class AmplifyOutput(
-    val value: Int,
+    val value: Long,
     val settings: Settings
 )
 
 // --- Single Pass Amplify ---
 
-fun amplify(program: Program, settings: Settings): Int {
-    fun nextOutput(input: Pair<Int, Int>) = runProgramWithInput(program, input.first, input.second).first()
+fun amplify(program: Program, settings: Settings): Long {
+    fun nextOutput(input: Pair<Long, Long>) = runProgramWithInput(program, input.first, input.second).first()
 
     val settingsIterator = (settings + 0).iterator()
-    return generateSequence(settingsIterator.next() to 0) {
+    return generateSequence(settingsIterator.next() to 0L) {
         if (settingsIterator.hasNext())
             settingsIterator.next() to nextOutput(it)
         else
@@ -35,7 +35,7 @@ fun amplify(program: Program, settings: Settings): Int {
 }
 
 fun optimalOutput(program: Program): AmplifyOutput? =
-    allPermutationsOf(listOf(0, 1, 2, 3, 4))
+    allPermutationsOf(listOf(0L, 1L, 2L, 3L, 4L))
         .map { settings ->
             AmplifyOutput(amplify(program, settings), settings)
         }
@@ -44,7 +44,7 @@ fun optimalOutput(program: Program): AmplifyOutput? =
 // --- Amplify with Feedback Loop ---
 
 private fun amplifyCluster(program: Program, settings: Settings): QueueComputerCluster {
-    fun queueWith(input: Int, nextDevice: QueueInputOutputDevice? = null) =
+    fun queueWith(input: Long, nextDevice: QueueInputOutputDevice? = null) =
         QueueInputOutputDevice(nextDevice).apply { putInput(input) }
 
     val linkedInputOutputDevices = settings.dropLast(1).reversed()
@@ -59,7 +59,7 @@ private fun amplifyCluster(program: Program, settings: Settings): QueueComputerC
     return QueueComputerCluster(computers)
 }
 
-fun amplifyWithFeedback(program: Program, settings: Settings): Int {
+fun amplifyWithFeedback(program: Program, settings: Settings): Long {
     val cluster = amplifyCluster(program, settings).apply { runAsync() }
 
     cluster.putInput(0)
@@ -75,7 +75,7 @@ fun amplifyWithFeedback(program: Program, settings: Settings): Int {
 }
 
 fun optimalOutputWithFeedback(program: Program): AmplifyOutput? =
-    allPermutationsOf(listOf(5, 6, 7, 8, 9))
+    allPermutationsOf(listOf(5L, 6L, 7L, 8L, 9L))
         .map { settings ->
             AmplifyOutput(amplifyWithFeedback(program, settings), settings)
         }
